fix conversations

+++ app/src/main/java/com/keylesspalace/tusky/components/conversation/ConversationsRemoteMediator.kt
@@ -4,8 +4,11 @@ import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
+import androidx.room.withTransaction
 import com.keylesspalace.tusky.db.AppDatabase
 import com.keylesspalace.tusky.network.MastodonApi
+import com.keylesspalace.tusky.util.HttpHeaderLink
+import retrofit2.HttpException
 
 @OptIn(ExperimentalPagingApi::class)
 class ConversationsRemoteMediator(
@@ -14,34 +17,54 @@ class ConversationsRemoteMediator(
     private val db: AppDatabase
 ) : RemoteMediator<Int, ConversationEntity>() {
 
+    private var nextKey: String? = null
+
+    private var order: Int = 0
+
     override suspend fun load(
         loadType: LoadType,
         state: PagingState<Int, ConversationEntity>
     ): MediatorResult {
 
         try {
-            val conversationsResult = when (loadType) {
+            val conversationsResponse = when (loadType) {
                 LoadType.REFRESH -> {
-                    api.getConversations(limit = state.config.initialLoadSize)
+                    nextKey = null
+                    order = 0
+                    api.getConversations(limit = state.config.pageSize)
                 }
                 LoadType.PREPEND -> {
                     return MediatorResult.Success(endOfPaginationReached = true)
                 }
                 LoadType.APPEND -> {
-                    val maxId = state.pages.findLast { it.data.isNotEmpty() }?.data?.lastOrNull()?.lastStatus?.id
-                    api.getConversations(maxId = maxId, limit = state.config.pageSize)
+                    api.getConversations(maxId = nextKey, limit = state.config.pageSize)
                 }
             }
 
-            if (loadType == LoadType.REFRESH) {
-                db.conversationDao().deleteForAccount(accountId)
+            val conversations = conversationsResponse.body()
+            if (!conversationsResponse.isSuccessful || conversations == null) {
+                return MediatorResult.Error(HttpException(conversationsResponse))
             }
-            db.conversationDao().insert(
-                conversationsResult
-                    .filterNot { it.lastStatus == null }
-                    .map { it.toEntity(accountId) }
-            )
-            return MediatorResult.Success(endOfPaginationReached = conversationsResult.isEmpty())
+
+            db.withTransaction {
+
+                if (loadType == LoadType.REFRESH) {
+                    db.conversationDao().deleteForAccount(accountId)
+                }
+
+                val linkHeader = conversationsResponse.headers()["Link"]
+                val links = HttpHeaderLink.parse(linkHeader)
+                nextKey = HttpHeaderLink.findByRelationType(links, "next")?.uri?.getQueryParameter("max_id")
+
+                db.conversationDao().insert(
+                    conversations
+                        .filterNot { it.lastStatus == null }
+                        .map {
+                            it.toEntity(accountId, order++)
+                        }
+                )
+            }
+            return MediatorResult.Success(endOfPaginationReached = nextKey == null)
         } catch (e: Exception) {
             return MediatorResult.Error(e)
         }
